Removed all c++ data structure
If someone need to use c++ features than is useless to try compile the codebase in c mode
because the user cannot include use cpp header and source anyways. It is better to use compile
the codebase with in c++ mode because it also can compile c files using `export "C"`.